fix 1.0-SNAPSHOT__Basic_schema.sql

+++ labs-oop/src/main/resources/db/sql/migration/1.0-SNAPSHOT__Basic_schema.sql
@@ -4,14 +4,14 @@ create table labs.t_function(
     id serial primary key,
     c_function_type varchar(255),
     c_count integer check ( c_count >= 2 ),
-    c_x_from double,
-    c_x_to double
+    c_x_from double precision,
+    c_x_to double precision
 );
 
 create table labs.t_point(
     id serial primary key,
-    function_id int,
-    c_x_value double,
-    c_y_value double,
+    function_id integer,
+    c_x_value double precision,
+    c_y_value double precision,
     foreign key (function_id) references labs.t_math_function(id) on delete cascade
 );
